challenge/requirements.txt
Django==4.1.1
django-cors-headers==3.13.0
django-dbbackup==4.0.2
djangorestframework==3.14.0
pip==22.2.2
pytz==2022.4
setuptools==65.3.0
sqlparse==0.4.3
asgiref==3.5.2